+++ ios/Classes/FlutterJlOtaPlugin.m
@@ -18,12 +18,19 @@
 - (void)handleMethodCall:(FlutterMethodCall *)call result:(FlutterResult)result {
     if ([@"getPlatformVersion" isEqualToString:call.method]) {
         result([@"iOS " stringByAppendingString:[[UIDevice currentDevice] systemVersion]]);
+    } else if ([@"getSdkVersion" isEqualToString:call.method]) {
+        result([[OtaTool sharedInstance] sdkVersion]);
+    } else if ([@"isOta" isEqualToString:call.method]) {
+        result(@([[OtaTool sharedInstance] isOtaUpdateInProgress]));
     } else if ([@"startScan" isEqualToString:call.method]) {
         [[OtaTool sharedInstance] startScan];
         result(@YES);
+    } else if ([@"stopScan" isEqualToString:call.method]) {
+        [[OtaTool sharedInstance] stopScan];
+        result(@YES);
     } else if ([@"connectDevice" isEqualToString:call.method]) {
         NSString *uuid = call.arguments[@"uuid"];
-        if ([uuid isKindOfClass:[NSString class]]) {
+        if ([uuid isKindOfClass:[NSString class]] && uuid.length > 0) {
             [[OtaTool sharedInstance] connectDeviceWithUUID:uuid];
             result(@YES);
         } else {
@@ -38,18 +45,20 @@
         if ([params isKindOfClass:[NSDictionary class]]) {
             NSString *uuid = params[@"uuid"];
             NSString *filePath = params[@"filePath"];
-            if ([uuid isKindOfClass:[NSString class]] && [filePath isKindOfClass:[NSString class]]) {
-                [[OtaTool sharedInstance] startOtaWithUuid:uuid filePath:filePath];
-
+            if ([uuid isKindOfClass:[NSString class]] && uuid.length > 0 &&
+                [filePath isKindOfClass:[NSString class]] && filePath.length > 0) {
                 __weak typeof(self) weakSelf = self;
                 [[OtaTool sharedInstance] setOtaProgressCallback:^(NSInteger progress, NSString *status) {
                     NSDictionary *response = @{
                             @"progress": @(progress),
                             @"status": status ?: @""
                     };
-                    [weakSelf.channel invokeMethod:@"otaProgress" arguments:response];
+                    dispatch_async(dispatch_get_main_queue(), ^{
+                        [weakSelf.channel invokeMethod:@"otaProgress" arguments:response];
+                    });
                 }];
 
+                [[OtaTool sharedInstance] startOtaWithUuid:uuid filePath:filePath];
                 result(@YES);
             } else {
                 result([FlutterError errorWithCode:@"INVALID_PARAMS"