+++ ios/Framework/JL_BLEKit.framework/Headers/JL_Assist.h
@@ -9,6 +9,13 @@
 #import <Foundation/Foundation.h>
 #import <CoreBluetooth/CoreBluetooth.h>
 
+
+@protocol JL_AssistDelegate <NSObject>
+
+-(void)assistDidWriteData:(NSData *_Nonnull)data;
+
+@end
+
 NS_ASSUME_NONNULL_BEGIN
 typedef void(^JL_Assist_BK)(BOOL isPaired);
 
@@ -23,6 +30,14 @@ typedef void(^JL_Assist_BK)(BOOL isPaired);
 @property(assign,nonatomic)BOOL               mNeedPaired;    //是否需要配对
 @property(assign,nonatomic)BOOL               mLogData;       //是否打印裸数据
 @property(assign,nonatomic)NSInteger          mLimitMtu;      //(默认40)在最大的MTU基础上减少数据量
+@property(strong,nonatomic)CBPeripheral       *__nullable mRcspPeripheral;
+@property(strong,nonatomic)CBCharacteristic   *__nullable mRcspWrite;
+@property(strong,nonatomic)CBCharacteristic   *__nullable mRcspRead;
+@property(strong,nonatomic)NSString           *mBleName;      //设备名字
+
+///代理,如果实现了此代理则发出的数据都需要走回调发送，内部不处理
+/// @discussion JL_AssistDelegate
+@property(assign,nonatomic)id<JL_AssistDelegate> __nullable mDelegate;
 
 /// Execute in a method 「- (void)centralManagerDidUpdateState:」
 /// @param state CBManagerState
@@ -51,6 +66,10 @@ typedef void(^JL_Assist_BK)(BOOL isPaired);
 /// @param characteristic CBCharacteristic
 -(void)assistUpdateValueForCharacteristic:(CBCharacteristic *)characteristic;
 
+
+/// Execute in a method 「- (void)peripheral:didWriteValueForCharacteristic:error:」
+/// Execute in a method 「- (void)peripheral:didIsReadyForWrite:error:」
+-(void)assistDidReady;
 
 @end