+++ ios/Framework/JL_BLEKit.framework/Headers/JL_OpCode.h
@@ -81,6 +81,7 @@
 #define kJL_TWS_NAME_LIST           0x0031   //一拖二/获取设备已连接手机名
 #define kJL_AI_CONTROL              0x0032   //AI控制命令
 #define kJL_PUBLIC_SET              0x0033   //公共设置命令
+#define kJL_TRAMSLATE               0x0034   //翻译命令
 
 
 #define kJL_WEAR_REQ_OPCODE         0x00A0   //获取运动信息